+++ dao-orm/src/main/java/com/training/carsharing/repository/ModelRepository.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository;
 
 import com.training.carsharing.model.impl.Model;
 
-public interface ModelRepository extends AbstractRepository<Model, Integer> {
+public interface ModelRepository extends AbstractRepository<Model, Long> {
 
 }
