+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/BrandRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Brand;
 
-public interface BrandRepositoryCustom extends AbstractRepositoryCustom<Brand, Integer> {
+public interface BrandRepositoryCustom extends AbstractRepositoryCustom<Brand, Long> {
 
 }
