+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/CalendarRepositoryCustom.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Calendar;
 
-public interface CalendarRepositoryCustom extends AbstractRepositoryCustom<Calendar, Integer> {
+public interface CalendarRepositoryCustom extends AbstractRepositoryCustom<Calendar, Long> {
 }
