+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/DriveRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Drive;
 
-public interface DriveRepositoryCustom extends AbstractRepositoryCustom<Drive, Integer> {
+public interface DriveRepositoryCustom extends AbstractRepositoryCustom<Drive, Long> {
 
 }
