+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/FuelRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Fuel;
 
-public interface FuelRepositoryCustom extends AbstractRepositoryCustom<Fuel, Integer> {
+public interface FuelRepositoryCustom extends AbstractRepositoryCustom<Fuel, Long> {
 
 }
