+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/MessageRepositoryCustom.java
@@ -2,5 +2,5 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Message;
 
-public interface MessageRepositoryCustom extends AbstractRepositoryCustom<Message, Integer> {
+public interface MessageRepositoryCustom extends AbstractRepositoryCustom<Message, Long> {
 }
