+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/ModelRepositoryCustom.java
@@ -2,6 +2,6 @@ package com.training.carsharing.repository.customrepository;
 
 import com.training.carsharing.model.impl.Model;
 
-public interface ModelRepositoryCustom extends AbstractRepositoryCustom<Model, Integer> {
+public interface ModelRepositoryCustom extends AbstractRepositoryCustom<Model, Long> {
 
 }
