+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/BodyTypeRepositoryImpl.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 
 @Repository
-public class BodyTypeRepositoryImpl extends AbstractRepositoryImpl<BodyType, Integer> implements BodyTypeRepositoryCustom {
+public class BodyTypeRepositoryImpl extends AbstractRepositoryImpl<BodyType, Long> implements BodyTypeRepositoryCustom {
 
     protected BodyTypeRepositoryImpl() {
         super(BodyType.class);
@@ -29,7 +29,7 @@ public class BodyTypeRepositoryImpl extends AbstractRepositoryImpl<BodyType, Int
 
     //TODO delete methods
     @Override
-    public BodyType findOneFullInfo(final Integer id) {
+    public BodyType findOneFullInfo(final Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<BodyType> cq = cb.createQuery(BodyType.class);
