+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/BrandRepositoryImpl.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 
 @Repository
-public class BrandRepositoryImpl extends AbstractRepositoryImpl<Brand, Integer> implements BrandRepositoryCustom {
+public class BrandRepositoryImpl extends AbstractRepositoryImpl<Brand, Long> implements BrandRepositoryCustom {
 
     protected BrandRepositoryImpl() {
         super(Brand.class);
@@ -29,7 +29,7 @@ public class BrandRepositoryImpl extends AbstractRepositoryImpl<Brand, Integer> 
 
     //TODO delete methods
     @Override
-    public Brand findOneFullInfo(final Integer id) {
+    public Brand findOneFullInfo(final Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<Brand> cq = cb.createQuery(Brand.class);
