+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/CalendarRepositoryImpl.java
@@ -15,7 +15,7 @@ import java.util.List;
 
 
 @Repository
-public class CalendarRepositoryImpl extends AbstractRepositoryImpl<Calendar, Integer> implements CalendarRepositoryCustom {
+public class CalendarRepositoryImpl extends AbstractRepositoryImpl<Calendar, Long> implements CalendarRepositoryCustom {
 
     protected CalendarRepositoryImpl() {
         super(Calendar.class);
@@ -29,7 +29,7 @@ public class CalendarRepositoryImpl extends AbstractRepositoryImpl<Calendar, Int
     }
 
     @Override
-    public Calendar findOneFullInfo(final Integer id) {
+    public Calendar findOneFullInfo(final Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<Calendar> cq = cb.createQuery(Calendar.class);
