+++ dao-orm/src/main/java/com/training/carsharing/repository/customrepository/impl/GearboxRepositoryImpl.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 
 @Repository
-public class GearboxRepositoryImpl extends AbstractRepositoryImpl<Gearbox, Integer> implements GearboxRepositoryCustom {
+public class GearboxRepositoryImpl extends AbstractRepositoryImpl<Gearbox, Long> implements GearboxRepositoryCustom {
 
     protected GearboxRepositoryImpl() {
         super(Gearbox.class);
@@ -29,7 +29,7 @@ public class GearboxRepositoryImpl extends AbstractRepositoryImpl<Gearbox, Integ
 
     //TODO delete methods
     @Override
-    public Gearbox findOneFullInfo(final Integer id) {
+    public Gearbox findOneFullInfo(final Long id) {
         final EntityManager em = getEntityManager();
         final CriteriaBuilder cb = em.getCriteriaBuilder();
         final CriteriaQuery<Gearbox> cq = cb.createQuery(Gearbox.class);
