+++ service/src/main/java/com/training/carsharing/AdService.java
@@ -3,6 +3,6 @@ package com.training.carsharing;
 import com.training.carsharing.model.impl.Ad;
 import com.training.carsharing.repository.AdRepository;
 
-public interface AdService extends AbstractService<Ad, Integer> {
+public interface AdService extends AbstractService<Ad, Long> {
 
 }
