+++ service/src/main/java/com/training/carsharing/BrandService.java
@@ -3,7 +3,7 @@ package com.training.carsharing;
 import com.training.carsharing.model.impl.Brand;
 import com.training.carsharing.repository.BrandRepository;
 
-public interface BrandService extends AbstractService<Brand, Integer> {
+public interface BrandService extends AbstractService<Brand, Long> {
 
 }
