+++ service/src/main/java/com/training/carsharing/DriveService.java
@@ -3,7 +3,7 @@ package com.training.carsharing;
 import com.training.carsharing.model.impl.Drive;
 import com.training.carsharing.repository.DriveRepository;
 
-public interface DriveService extends AbstractService<Drive, Integer> {
+public interface DriveService extends AbstractService<Drive, Long> {
 
 }
