+++ service/src/main/java/com/training/carsharing/GearboxService.java
@@ -3,7 +3,7 @@ package com.training.carsharing;
 import com.training.carsharing.model.impl.Gearbox;
 import com.training.carsharing.repository.GearboxRepository;
 
-public interface GearboxService extends AbstractService<Gearbox, Integer> {
+public interface GearboxService extends AbstractService<Gearbox, Long> {
 
 }
