+++ service/src/main/java/com/training/carsharing/PassportService.java
@@ -3,7 +3,7 @@ package com.training.carsharing;
 import com.training.carsharing.model.impl.Passport;
 import com.training.carsharing.repository.PassportRepository;
 
-public interface PassportService extends AbstractService<Passport, Integer> {
+public interface PassportService extends AbstractService<Passport, Long> {
 
 }
