+++ service/src/main/java/com/training/carsharing/impl/CustomBodyTypeService.java
@@ -6,7 +6,7 @@ import com.training.carsharing.repository.BodyTypeRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CustomBodyTypeService extends CustomAbstractService<BodyType, Integer> implements BodyTypeService {
+public class CustomBodyTypeService extends CustomAbstractService<BodyType, Long> implements BodyTypeService {
 
     public CustomBodyTypeService() { super(BodyType.class); }
