+++ service/src/main/java/com/training/carsharing/impl/CustomBrandService.java
@@ -6,7 +6,7 @@ import com.training.carsharing.repository.BrandRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CustomBrandService extends CustomAbstractService<Brand, Integer> implements BrandService {
+public class CustomBrandService extends CustomAbstractService<Brand, Long> implements BrandService {
 
     public CustomBrandService() { super(Brand.class);
     }
