+++ service/src/main/java/com/training/carsharing/impl/CustomDriveService.java
@@ -6,7 +6,7 @@ import com.training.carsharing.repository.DriveRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CustomDriveService extends CustomAbstractService<Drive, Integer> implements DriveService {
+public class CustomDriveService extends CustomAbstractService<Drive, Long> implements DriveService {
 
     public CustomDriveService() { super(Drive.class); }
